Add DynamicDependency attributes where needed for trimming
